requirements.txt
Pillow~=7.0.0
numpy>=1.20.3
matplotlib>=3.4.2